backend/requirements.txt
# Data processing
pandas==2.2.2
numpy==1.26.4

# ML & embeddings
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
sentence-transformers==2.7.0
faiss-cpu==1.7.4

# LLM & RAG
langchain==0.1.20
langchain-groq==0.1.6

# API server
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# Utilities
python-dotenv==1.0.1
slowapi==0.1.9

# Auth
python-jose==3.3.0
passlib[bcrypt]==1.7.4